Make the $atts param mixed to handle empty string
When a user adds no attributes to the shortcode, an empty string is passed in the $atts value, throwing an error when $atts is expected to be an array.

This commit switches the type for $atts to be mixed, and then checks if the value is an empty string, and if so, convert to an empty array.

If the value is already an array, all is good.